Move AnimationState to script_traits
Resolves #9223
