+++ components/compositing/compositor_thread.rs
@@ -13,11 +13,10 @@ use headless;
 use ipc_channel::ipc::{self, IpcReceiver, IpcSender};
 use layers::layers::{BufferRequest, LayerBufferSet};
 use layers::platform::surface::{NativeDisplay, NativeSurface};
-use msg::constellation_msg::{AnimationState, PipelineId};
-use msg::constellation_msg::{Image, Key, KeyModifiers, KeyState};
+use msg::constellation_msg::{Image, Key, KeyModifiers, KeyState, PipelineId};
 use profile_traits::mem;
 use profile_traits::time;
-use script_traits::{EventResult, ScriptToCompositorMsg};
+use script_traits::{AnimationState, EventResult, ScriptToCompositorMsg};
 use std::fmt::{Debug, Error, Formatter};
 use std::rc::Rc;
 use std::sync::mpsc::{Receiver, Sender, channel};
