+++ components/compositing/headless.rs
@@ -7,10 +7,10 @@ use compositor_thread::{CompositorEventListener, CompositorReceiver};
 use compositor_thread::{InitialCompositorState, Msg};
 use euclid::scale_factor::ScaleFactor;
 use euclid::{Point2D, Size2D};
-use msg::constellation_msg::AnimationState;
 use msg::constellation_msg::WindowSizeData;
 use profile_traits::mem;
 use profile_traits::time;
+use script_traits::AnimationState;
 use std::sync::mpsc::Sender;
 use util::opts;
 use windowing::WindowEvent;
